订单分页查询

+++ sky-server/src/main/java/com/sky/service/impl/DishServiceImpl.java
@@ -2,19 +2,24 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
-import com.sky.context.BaseContext;
+import com.sky.constant.MessageConstant;
+import com.sky.constant.StatusConstant;
 import com.sky.dto.DishDTO;
 import com.sky.dto.DishPageQueryDTO;
 import com.sky.entity.Dish;
+import com.sky.entity.DishFlavor;
 import com.sky.exception.CheckException;
+import com.sky.exception.DeletionNotAllowedException;
+import com.sky.mapper.DishFlavorMapper;
 import com.sky.mapper.DishMapper;
 import com.sky.result.PageResult;
 import com.sky.service.DishService;
 import com.sky.vo.DishVO;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
+import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
@@ -22,8 +27,12 @@ import java.util.Objects;
 public class DishServiceImpl implements DishService {
     private final DishMapper dishMapper;
 
-    public DishServiceImpl(DishMapper dishMapper) {
+    private final DishFlavorMapper dishFlavorMapper;
+    ;
+
+    public DishServiceImpl(DishMapper dishMapper, DishFlavorMapper dishFlavorMapper) {
         this.dishMapper = dishMapper;
+        this.dishFlavorMapper = dishFlavorMapper;
     }
 
     @Override
@@ -35,6 +44,7 @@ public class DishServiceImpl implements DishService {
         DishVO dish = new DishVO();
         dish.setCategoryId(dishPageQueryDTO.getCategoryId());
         dish.setName(dishPageQueryDTO.getName());
+        dish.setStatus(dishPageQueryDTO.getStatus());
         //原始查询
         List<DishVO> dishes = dishMapper.pageQuery(dish);
         PageResult pageResult = new PageResult();
@@ -43,6 +53,7 @@ public class DishServiceImpl implements DishService {
         pageResult.setRecords(dishPage.getResult());
         return pageResult;
     }
+
 
     @Override
     public void updateDish(DishDTO dishDTO) {
@@ -53,21 +64,40 @@ public class DishServiceImpl implements DishService {
         if (count > 0) {
             throw new CheckException("输入的菜品名称重复啦");
         }
-        dish.setUpdateTime(LocalDateTime.now());
-        dish.setUpdateUser(BaseContext.getCurrentId());
         dishMapper.updateDish(dish);
+        //更新口味
+        dishFlavorMapper.deleteByDishId(Collections.singletonList(dish.getId()));
+        List<DishFlavor> flavors = dishFlavorMapper.getDishFlavor(dish.getId());
+        if (Objects.nonNull(flavors) && !flavors.isEmpty()) {
+            flavors.forEach(item -> item.setDishId(dish.getId()));
+            dishFlavorMapper.insert(flavors);
 
+        }
 
     }
 
     @Override
     public DishVO getDish(Long id) {
-        return dishMapper.getDish(id);
+        Dish dish = dishMapper.dish(id);
+
+        DishVO dishVO = new DishVO();
+        List<DishFlavor> dishFlavors = dishFlavorMapper.getDishFlavor(id);
+
+        BeanUtils.copyProperties(dish, dishVO);
+        dishVO.setFlavors(dishFlavors);
+        return dishVO;
     }
 
     @Override
+    @Transactional
     public void deleteById(List<Long> ids) {
+        Integer count = dishMapper.selcetCountByIdsAndStatus(ids, StatusConstant.ENABLE);
+        if (count > 0) {
+            throw new DeletionNotAllowedException(MessageConstant.DISH_ON_SALE);
+        }
         dishMapper.deleteById(ids);
+        dishFlavorMapper.deleteByDishId(ids);
+
     }
 
     @Override
@@ -82,25 +112,30 @@ public class DishServiceImpl implements DishService {
         Dish dish = new Dish();
         dish.setId(id);
         dish.setStatus(status);
-        dish.setUpdateTime(LocalDateTime.now());
-        dish.setUpdateUser(BaseContext.getCurrentId());
+
         dishMapper.updateDish(dish);
+
+
     }
 
     @Override
+    @Transactional
     public void insertDish(DishVO dishVO) {
 
         Dish dish = new Dish();
         BeanUtils.copyProperties(dishVO, dish);
         Integer count = dishMapper.selcetCount(dish);
         if (count > 0) {
-            throw new CheckException("输入的菜品名称重复啦");
+            throw new CheckException(MessageConstant.DISH_CHECK);
         }
-        dish.setUpdateUser(BaseContext.getCurrentId());
-        dish.setCreateUser(BaseContext.getCurrentId());
-        dish.setCreateTime(LocalDateTime.now());
-        dish.setUpdateTime(LocalDateTime.now());
+
         dishMapper.insertDish(dish);
+        List<DishFlavor> flavors = dishVO.getFlavors();
+        if (Objects.nonNull(flavors) && !flavors.isEmpty()) {
+            flavors.forEach(item -> item.setDishId(dish.getId()));
+            dishFlavorMapper.insert(flavors);
+
+        }
     }
 
 }
